refactor: rewrite in WinUI 3
