--- App1/App1/MainWindow.xaml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<Page
-    x:Class="App1.MainWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:App1"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:tb="using:H.NotifyIcon"
-    mc:Ignorable="d">
-
-    <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center">
-        <Button x:Name="myButton" Click="myButton_Click">Click Me</Button>
-        <tb:TaskbarIcon
-            x:Name="NTaskbarIcon"
-            ToolTipText="ProcessReporter"
-            LeftClickCommand="{Binding ShowWindowCommand}"
-            NoLeftClickDelay="True">
-            <tb:TaskbarIcon.ContextFlyout>
-                <MenuFlyout>
-                    <MenuFlyoutItem Command="{Binding ShowWindowCommand}" Text="显示主程序" />
-                    <MenuFlyoutSeparator />
-                    <MenuFlyoutItem Command="{Binding ExitAppCommand}" Text="退出" />
-                </MenuFlyout>
-            </tb:TaskbarIcon.ContextFlyout>
-        </tb:TaskbarIcon>
-    </StackPanel>
-</Page>
