--- App1/App1/Views/TrayIconView.xaml
@@ -1,43 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-﻿<UserControl
-    x:Class="App1.Views.TrayIconView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:converters="using:App1.Converters"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:tb="using:H.NotifyIcon"
-    mc:Ignorable="d"
-    >
-    <UserControl.Resources>
-        <converters:AppThemeToImageSourceConverter
-            x:Key="AppThemeToImageSourceConverter"
-            LightImage="ms-appx:///Resources/AppIcon/flower.svg"
-            DarkImage="ms-appx:///Resources/flower_light.svg"
-            />
-    </UserControl.Resources>
-    
-    <tb:TaskbarIcon
-        x:Name="TrayIcon"
-        x:FieldModifier="public"
-        ContextMenuMode="SecondWindow"
-        IconSource="{x:Bind Theme, Converter={StaticResource AppThemeToImageSourceConverter}}"
-        LeftClickCommand="{x:Bind ShowHideWindowCommand}"
-        NoLeftClickDelay="True"
-        ToolTipText="ToolTip"
-        >
-        <tb:TaskbarIcon.ContextFlyout>
-            <MenuFlyout AreOpenCloseAnimationsEnabled="False">
-                <MenuFlyoutItem
-                    Command="{x:Bind ShowHideWindowCommand}"
-                    Text="Show/Hide Window"
-                    />
-                <MenuFlyoutSeparator />
-                <MenuFlyoutItem
-                    Command="{x:Bind ExitApplicationCommand}"
-                    Text="Exit"
-                    />
-            </MenuFlyout>
-        </tb:TaskbarIcon.ContextFlyout>
-    </tb:TaskbarIcon>
-</UserControl>
