--- ProcessReporterWin.sln
@@ -1,89 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.8.34525.116
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProcessReporterWin", "ProcessReporterWin\ProcessReporterWin.csproj", "{2AC50C5C-39B7-4E49-BECC-5331343EA959}"
-EndProject
-Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "App1 (Package)", "App1\App1 (Package)\App1 (Package).wapproj", "{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App1", "App1\App1\App1.csproj", "{E47A9748-474B-4379-97E1-5B7A6EAD9718}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|x64.Build.0 = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Debug|x86.Build.0 = Debug|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|ARM64.Build.0 = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|x64.ActiveCfg = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|x64.Build.0 = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|x86.ActiveCfg = Release|Any CPU
-		{2AC50C5C-39B7-4E49-BECC-5331343EA959}.Release|x86.Build.0 = Release|Any CPU
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|Any CPU.Build.0 = Debug|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|Any CPU.Deploy.0 = Debug|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|ARM64.Build.0 = Debug|ARM64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|x64.ActiveCfg = Debug|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|x64.Build.0 = Debug|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|x64.Deploy.0 = Debug|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|x86.ActiveCfg = Debug|x86
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|x86.Build.0 = Debug|x86
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Debug|x86.Deploy.0 = Debug|x86
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|Any CPU.ActiveCfg = Release|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|Any CPU.Build.0 = Release|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|Any CPU.Deploy.0 = Release|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|ARM64.ActiveCfg = Release|ARM64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|ARM64.Build.0 = Release|ARM64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|ARM64.Deploy.0 = Release|ARM64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|x64.ActiveCfg = Release|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|x64.Build.0 = Release|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|x64.Deploy.0 = Release|x64
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|x86.ActiveCfg = Release|x86
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|x86.Build.0 = Release|x86
-		{9AAEFB9C-1B6A-4633-87BC-588FFDF04B38}.Release|x86.Deploy.0 = Release|x86
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|Any CPU.Build.0 = Debug|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|ARM64.Build.0 = Debug|ARM64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|x64.ActiveCfg = Debug|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|x64.Build.0 = Debug|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|x86.ActiveCfg = Debug|x86
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Debug|x86.Build.0 = Debug|x86
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|Any CPU.ActiveCfg = Release|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|Any CPU.Build.0 = Release|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|ARM64.ActiveCfg = Release|ARM64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|ARM64.Build.0 = Release|ARM64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|x64.ActiveCfg = Release|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|x64.Build.0 = Release|x64
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|x86.ActiveCfg = Release|x86
-		{E47A9748-474B-4379-97E1-5B7A6EAD9718}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CE9C717E-F1E7-45B8-AA13-092DCE0C9B28}
-	EndGlobalSection
-EndGlobal
