+++ ProcessReporterWin/Layouts/Main.xaml
@@ -4,8 +4,7 @@
             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
             xmlns:layouts="using:ProcessReporterWin.Layouts"
             x:Class="ProcessReporterWin.Layouts.Main"
-            x:DataType="layouts:Main"
-            Padding="20">
+            x:DataType="layouts:Main">
     <FlyoutPage.Flyout>
         <layouts:SideBar x:Name="SideBar" />
     </FlyoutPage.Flyout>
