+++ ProcessReporterWin/ProcessReporterWin.csproj
@@ -1,21 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFrameworks>;net8.0-windows10.0.19041.0</TargetFrameworks>
-        <TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows'))">$(TargetFrameworks);net8.0-windows10.0.19041.0</TargetFrameworks>
-        <!-- Uncomment to also build the tizen app. You will need to install tizen by following this: https://github.com/Samsung/Tizen.NET -->
-        <!-- <TargetFrameworks>$(TargetFrameworks);net8.0-tizen</TargetFrameworks> -->
-
-        <!-- Note for MacCatalyst:
-        The default runtime is maccatalyst-x64, except in Release config, in which case the default is maccatalyst-x64;maccatalyst-arm64.
-        When specifying both architectures, use the plural <RuntimeIdentifiers> instead of the singular <RuntimeIdentifier>.
-        The Mac App Store will NOT accept apps with ONLY maccatalyst-arm64 indicated;
-        either BOTH runtimes must be indicated or ONLY macatalyst-x64. -->
-        <!-- For example: <RuntimeIdentifiers>maccatalyst-x64;maccatalyst-arm64</RuntimeIdentifiers> -->
-		
-        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0-windows10.0.20348.0</TargetFramework>
+        <TargetPlatformMinVersion>10.0.19041.0</TargetPlatformMinVersion>
+        <SupportedOSPlatformVersion>10.0.19041.0</SupportedOSPlatformVersion>
+        <OutputType>WinExe</OutputType>
         <RootNamespace>ProcessReporterWin</RootNamespace>
-        <UseMaui>true</UseMaui>
+        <ApplicationManifest>app.manifest</ApplicationManifest>
+        <UseWinUI>true</UseWinUI>
         <SingleProject>true</SingleProject>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
@@ -51,22 +43,35 @@
         <!-- Splash Screen -->
 
         <!-- Images -->
-        <MauiImage Include="Resources\Images\*"/>
+        <MauiImage Include="Resources\Images\*" />
 
         <!-- Custom Fonts -->
-        <MauiFont Include="Resources\Fonts\*"/>
+        <MauiFont Include="Resources\Fonts\*" />
 
         <!-- Raw Assets (also remove the "Resources\Raw" prefix) -->
-        <MauiAsset Include="Resources\Raw\**" LogicalName="%(RecursiveDir)%(Filename)%(Extension)"/>
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="CommunityToolkit.Mvvm" Version="8.2.2"/>
+      <Compile Remove="Extensions\**" />
+      <Compile Remove="Resources\Raw\**" />
+      <EmbeddedResource Remove="Extensions\**" />
+      <EmbeddedResource Remove="Resources\Raw\**" />
+      <MauiCss Remove="Extensions\**" />
+      <MauiCss Remove="Resources\Raw\**" />
+      <MauiXaml Remove="Extensions\**" />
+      <MauiXaml Remove="Resources\Raw\**" />
+      <None Remove="Extensions\**" />
+      <None Remove="Resources\Raw\**" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="CommunityToolkit.Mvvm" Version="8.2.2" />
+        <PackageReference Include="CommunityToolkit.WinUI.UI.Controls" Version="7.1.2" />
         <PackageReference Include="H.NotifyIcon.Maui" Version="2.0.131" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
         <PackageReference Include="Microsoft.Maui.Controls" Version="8.0.21" />
         <PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="8.0.21" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0" />
         <PackageReference Include="Microsoft.Windows.CsWin32" Version="0.3.49-beta">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -75,16 +80,22 @@
     </ItemGroup>
 
     <ItemGroup>
-        <Folder Include="Extensions\"/>
         <Folder Include="Pages\" />
-        <Folder Include="Resources\Images\"/>
-        <Folder Include="Resources\Raw\"/>
+        <Folder Include="Resources\Images\" />
         <Folder Include="Resources\Splash\" />
     </ItemGroup>
 
     <ItemGroup>
       <MauiIcon Include="Resources\AppIcon\flower.svg" />
       <MauiIcon Include="Resources\AppIcon\flower_light.svg" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PRIResource Remove="Resources\Raw\**" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PRIResource Remove="Extensions\**" />
     </ItemGroup>
 
 </Project>
