+++ ShiroProcessReporter/ShiroProcessReporter (Package)/Images/LockScreenLogo.scale-200.png
