+++ ShiroProcessReporter/ShiroProcessReporter (Package)/Images/SplashScreen.scale-200.png
