+++ ShiroProcessReporter/ShiroProcessReporter (Package)/Images/Square150x150Logo.scale-200.png
