+++ ShiroProcessReporter/ShiroProcessReporter (Package)/Images/Square44x44Logo.scale-200.png
