+++ ShiroProcessReporter/ShiroProcessReporter (Package)/Images/Square44x44Logo.targetsize-24_altform-unplated.png
