+++ ShiroProcessReporter/ShiroProcessReporter (Package)/Images/Wide310x150Logo.scale-200.png
