+++ ShiroProcessReporter/ShiroProcessReporter/App.xaml.cs
@@ -17,38 +17,18 @@ using Windows.ApplicationModel.Activation;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
 
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
-namespace App1
+namespace ShiroProcessReporter
 {
-    /// <summary>
-    /// Provides application-specific behavior to supplement the default Application class.
-    /// </summary>
     public partial class App : Application
     {
-        /// <summary>
-        /// Initializes the singleton application object.  This is the first line of authored code
-        /// executed, and as such is the logical equivalent of main() or WinMain().
-        /// </summary>
         public App()
         {
             this.InitializeComponent();
         }
 
-        /// <summary>
-        /// Invoked when the application is launched.
-        /// </summary>
-        /// <param name="args">Details about the launch request and process.</param>
         protected override void OnLaunched(Microsoft.UI.Xaml.LaunchActivatedEventArgs args)
         {
-            MainWindow = new Window
-            {
-                Content = new Frame
-                {
-                    Content = new MainWindow()
-                }
-            };
+            MainWindow = new MainView();
 
             MainWindow.Closed += (sender, args) =>
             {
@@ -59,6 +39,6 @@ namespace App1
             MainWindow.Activate();
         }
 
-        public static Window MainWindow { get; private set; }
+        public static Window? MainWindow { get; private set; }
     }
 }
