+++ ShiroProcessReporter/ShiroProcessReporter/Components/TrayIconView.xaml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="ShiroProcessReporter.Components.TrayIconView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:ShiroProcessReporter.Components"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:tb="using:H.NotifyIcon"
+    xmlns:converters="using:ShiroProcessReporter.Converters"
+    d:DataContext="{d:DesignInstance Type=local:TrayIconView}"
+    mc:Ignorable="d">
+
+    <UserControl.Resources>
+        <converters:AppThemeToImageSourceConverter
+            x:Key="AppThemeToImageSourceConverter"
+            LightImage="ms-appx:///Resources/TrayIcon/flower.ico"
+            DarkImage="ms-appx:///Resources/TrayIcon/flower_light.ico" />
+    </UserControl.Resources>
+
+    <tb:TaskbarIcon
+        x:Name="TrayIcon"
+        x:FieldModifier="public"
+        ContextMenuMode="SecondWindow"
+        IconSource="{x:Bind Theme, Converter={StaticResource AppThemeToImageSourceConverter}, Mode=OneWay}"
+        LeftClickCommand="{x:Bind ShowHideWindowCommand}"
+        NoLeftClickDelay="True"
+        ToolTipText="ShiroProcessReporter">
+        <tb:TaskbarIcon.ContextFlyout>
+            <MenuFlyout>
+                <MenuFlyoutItem Command="{x:Bind ShowHideWindowCommand}" Text="显示/隐藏设置" />
+                <MenuFlyoutSeparator />
+                <MenuFlyoutItem Command="{x:Bind ExitApplicationCommand}" Text="退出" />
+            </MenuFlyout>
+        </tb:TaskbarIcon.ContextFlyout>
+    </tb:TaskbarIcon>
+</UserControl>
