+++ ShiroProcessReporter/ShiroProcessReporter/Layouts/Navigation.xaml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="ShiroProcessReporter.Layouts.Navigation"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:ShiroProcessReporter.Layouts"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+    <Grid>
+        <NavigationView Background="Transparent" IsBackEnabled="False" IsBackButtonVisible="Collapsed">
+            <NavigationView.Header>
+                <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center">
+                    <Image Source="mx-apps://Resources/Images/mix_space.png" Width="300" Height="200" />
+                </StackPanel>
+            </NavigationView.Header>
+            <NavigationView.MenuItems>
+                <NavigationViewItem Content="Mix Space Endpoint">
+
+                </NavigationViewItem>
+                <NavigationViewItem Content="Replace Rules">
+
+                </NavigationViewItem>
+                <NavigationViewItem Content="Filter Rules">
+
+                </NavigationViewItem>
+            </NavigationView.MenuItems>
+        </NavigationView>
+    </Grid>
+</Page>
