+++ ShiroProcessReporter/ShiroProcessReporter/MainView.xaml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Window
+    x:Class="ShiroProcessReporter.MainView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:ShiroProcessReporter"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:components="using:ShiroProcessReporter.Components"
+    xmlns:layouts="using:ShiroProcessReporter.Layouts"
+    mc:Ignorable="d">
+
+    <Grid
+        VerticalAlignment="Stretch"
+        HorizontalAlignment="Stretch"
+        Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <layouts:Navigation />
+        <components:TrayIconView />
+    </Grid>
+</Window>
