+++ ShiroProcessReporter/ShiroProcessReporter/Resources/Images/mix_space.png
